# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Styling with Muskan API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/users/me": {
      "get": { "summary": "Get current user", "responses": { "200": { "description": "User data" }, "401": { "description": "Unauthorized" } } },
      "patch": {
        "summary": "Update profile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "name": { "type": "string" }, "referralCode": { "type": "string" } } }
            }
          }
        },
        "responses": { "200": { "description": "Updated" }, "400": { "description": "Validation errors" } }
      }
    },
    "/users/me/addresses": {
      "get": { "summary": "List addresses", "responses": { "200": { "description": "OK" } } },
      "post": {
        "summary": "Add address",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "houseNo": { "type": "string" }, "area": { "type": "string" }, "landmark": { "type": "string" }, "type": { "type": "string" }, "lat": { "type": "number" }, "lng": { "type": "number" } },
                "required": ["houseNo", "area"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Validation errors" } }
      }
    },
    "/users/me/addresses/{id}": {
      "delete": {
        "summary": "Delete address",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
...
    "/payments/razorpay/order": {
      "post": {
        "summary": "Create Razorpay order",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "amount": { "type": "integer" }, "currency": { "type": "string" }, "receipt": { "type": "string" } }, "required": ["amount"] } } }
        },
        "responses": { "200": { "description": "Order" }, "400": { "description": "Validation errors" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/payments/razorpay/verify": {
      "post": {
        "summary": "Verify Razorpay signature",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "order_id": { "type": "string" }, "payment_id": { "type": "string" }, "signature": { "type": "string" }, "bookingId": { "type": "string" }, "amount": { "type": "integer" } }, "required": ["order_id", "payment_id", "signature"] } } }
        },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Invalid signature" }, "401": { "description": "Unauthorized" } }
      }
    },
...
  }
}